refactor: clarify migration module names
